+++ stories/Post/100b-CompactPostThread.stories.tsx
@@ -1,14 +1,14 @@
 import CompactPostThread from "post/CompactPostThread";
 import React from "react";
 import { action } from "@storybook/addon-actions";
-import greedlerAvatar from "./images/greedler.jpg";
-import hannibalAvatar from "./images/hannibal.png";
-import mamoruAvatar from "./images/mamoru.png";
-import oncelerAvatar from "./images/oncie.jpg";
-import tuxedoAvatar from "./images/tuxedo-mask.jpg";
+import greedlerAvatar from "stories/images/greedler.jpg";
+import reindeerEars from "stories/images/reindeer-ears.png";
+import mamoruAvatar from "stories/images/mamoru.png";
+import oncelerAvatar from "stories/images/oncie.jpg";
+import tuxedoAvatar from "stories/images/tuxedo-mask.jpg";
 
 export default {
-  title: "Post Thread Preview",
+  title: "Posts / Compact Post Thread",
   component: CompactPostThread,
 };
 export const CompactThreadStory = () => {
